+++ src/www/admin/services/delete.php
@@ -5,7 +5,7 @@ use Garradin\Services\Services;
 
 require_once __DIR__ . '/_inc.php';
 
-$session->requireAccess('membres', Membres::DROIT_ADMIN);
+$session->requireAccess($session::SECTION_USERS, $session::ACCESS_ADMIN);
 
 $service = Services::get((int) qg('id'));
 
@@ -16,6 +16,10 @@ if (!$service) {
 $csrf_key = 'service_delete_' . $service->id();
 
 $form->runIf('delete', function () use ($service) {
+	if (!f('confirm_delete')) {
+		throw new UserException('Merci de cocher la case pour confirmer la suppression.');
+	}
+
 	$service->delete();
 }, $csrf_key, ADMIN_URL . 'services/');
