+++ src/www/admin/services/fees/delete.php
@@ -5,7 +5,7 @@ use Garradin\Services\Fees;
 
 require_once __DIR__ . '/../_inc.php';
 
-$session->requireAccess('membres', Membres::DROIT_ADMIN);
+$session->requireAccess($session::SECTION_USERS, $session::ACCESS_ADMIN);
 
 $fee = Fees::get((int) qg('id'));
 
@@ -16,6 +16,10 @@ if (!$fee) {
 $csrf_key = 'fee_delete_' . $fee->id();
 
 $form->runIf('delete', function () use ($fee) {
+	if (!f('confirm_delete')) {
+		throw new UserException('Merci de cocher la case pour confirmer la suppression.');
+	}
+
 	$fee->delete();
 }, $csrf_key, ADMIN_URL . 'services/fees/?id=' . $fee->id_service);
